+++ changelog.md
@@ -6,6 +6,8 @@
   * Has now 2 levels, increasing durability by 30/100 per level (on armor/tools)
   * Is now affected by durability modifiers
     * E.g. hammers bonus durability will be 133 instead of 400
+* Juicy bait reworked
+  * Now max 1 level, increases chance to double fishing loot by 15%
 * Same for Haste
   * Is now two levels, increasing mining speed by 1.5 per level
 * Reduced Explosive enchantment cost